Update 0029-semantic-versioning.md with clarification
This adds a clarification around removing dependencies. I have talked with folks about this, and we have done this for a long time on the Java team.

It came up recently, and I just wanted to try and synchronize this document with what has effectively been the way we've been bumping versions so there are no surprises.